==> notebooks/04_transform_raw_data_into_feature_and_targets.ipynb <==
#jupyter magic command
%reload_ext autoreload
%autoreload 2

print(os.getcwd())  # Check current working directory


import sys
import os

# Change the current working directory to the root of the project
os.chdir(os.path.abspath(os.path.join(os.getcwd(), '..')))  # Move up to the root directory

# Add the src directory to the Python path
sys.path.append(os.path.join(os.getcwd(), 'src'))

print(os.getcwd())  # Verify the new working directory
print(sys.path)  # Verify that 'src' is in the Python path

from src.data import load_raw_data

rides = load_raw_data(year=2024)

rides


from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*28*1, # one month
    step_size=24,
)

print(f'{features.shape=}')
print(f'{targets.shape=}')

# save back to disk
tabular_data = features
tabular_data['target_rides_next_hour'] = targets

from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
