==> notebooks/05_visualize_training_data.ipynb <==
import pandas as pd
import sys
import os

# Change the current working directory to the root of the project
os.chdir(os.path.abspath(os.path.join(os.getcwd(), '..')))  # Move up to the root directory

# Add the src directory to the Python path
sys.path.append(os.path.join(os.getcwd(), 'src'))

print(os.getcwd())  # Verify the new working directory
print(sys.path)  # Verify that 'src' is in the Python path

from src.paths import TRANSFORMED_DATA_DIR

features_and_target = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
features_and_target

features = features_and_target.drop(columns=['target_rides_next_hour'])
targets = features_and_target['target_rides_next_hour']

# Change the current working directory to the root of the project
os.chdir(os.path.abspath(os.path.join(os.getcwd(), '..')))  # Move up to the root directory

# Add the src directory to the Python path
sys.path.append(os.path.join(os.getcwd(), 'src'))

print(os.getcwd())  # Verify the new working directory
print(sys.path)  # Verify that 'src' is in the Python path
from src.plot import plot_one_sample

plot_one_sample(
    example_id=0,
    features=features,
    targets=targets,
)
